+++ NSM/train/deprecated/train_deep_sdf_orig.py
@@ -107,7 +107,6 @@ def train_deep_sdf(config, model, sdf_dataset, use_wandb=False):
                     convergence_patience=config["convergence_patience_recon"],
                     verbose=config["verbose"],
                     objects_per_decoder=1,
-                    batch_size_latent_recon=config["batch_size_latent_recon"],
                     get_rand_pts=config["get_rand_pts_recon"],
                     n_pts_random=config["n_pts_random_recon"],
                     sigma_rand_pts=config["sigma_rand_pts_recon"],